drm/amdgpu: extend ras feature mask to 64-bit for A + A
Extend ras feature mask to 64-bit for A + A.

+++ drivers/gpu/drm/amd/amdgpu/amdgpu.h
@@ -208,7 +208,7 @@ extern int amdgpu_backlight;
 extern int amdgpu_damage_clips;
 extern struct amdgpu_mgpu_info mgpu_info;
 extern int amdgpu_ras_enable;
-extern uint amdgpu_ras_mask;
+extern u64 amdgpu_ras_mask;
 extern int amdgpu_bad_page_threshold;
 extern bool amdgpu_ignore_bad_page_threshold;
 extern struct amdgpu_watchdog_timer amdgpu_watchdog_timer;
